Refactor code in Denied user component
