+++ src/constants/deniedUsersAction.js
@@ -1,0 +1,5 @@
+constants = {
+    ADD_TO_EVENT: 'Add to event'
+}
+
+export default constants;
