+++ src/containers/DeniedUsersAction/DeniedUsers.js
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+import { promoteToOwner, approveUser } from '../../actions/event/event-item-view-action';
+import DeniedUsersAction from '../../components/event/DeniedUsersAction/Denied-users-action';
+
+const mapStateToProps = (state) => ({
+  eventId: state.event.data.id
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  approveUser: (userId, eventId, buttonAction) => dispatch(approveUser(userId, eventId, buttonAction)),
+  promoteToOwner: (userId, eventId) => dispatch(promoteToOwner(userId, eventId))
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(DeniedUsersAction);
